+++ phpunit.xml.dist
@@ -2,21 +2,17 @@
 <phpunit colors="true">
     <ini name="date.timezone" value="UTC"/>
     <testsuites>
-<!--        <testsuite name="Community">-->
-<!--            <directory suffix="Test.php">tests/Integration/</directory>-->
-<!--            <exclude>./tests/Integration/KeysTest.php</exclude>-->
-<!--        </testsuite>-->
-<!--        <testsuite name="AlgoliaSearch">-->
-<!--            <directory suffix="Test.php">tests/Integration/</directory>-->
-<!--        </testsuite>-->
-<!--        <testsuite name="Definition">-->
-<!--            <file>tests/API/PublicApiTest.php</file>-->
-<!--        </testsuite>-->
-<!--        <testsuite name="Unit">-->
-<!--            <directory suffix="Test.php">tests/Unit/</directory>-->
-<!--        </testsuite>-->
-        <testsuite name="CtsIntegration">
-            <directory suffix="Test.php">tests/Cts/Integration</directory>
+        <testsuite name="Community">
+            <directory suffix="Test.php">tests/Integration/</directory>
+        </testsuite>
+        <testsuite name="AlgoliaSearch">
+            <directory suffix="Test.php">tests/Integration/</directory>
+        </testsuite>
+        <testsuite name="Definition">
+            <file>tests/API/PublicApiTest.php</file>
+        </testsuite>
+        <testsuite name="Unit">
+            <directory suffix="Test.php">tests/Unit/</directory>
         </testsuite>
     </testsuites>
 </phpunit>
